Deprecate and ignore the --backlog option
When I ported socket-burst-dampener to go and rust, I found that there
was no adjustable backlog setting for socket listeners exposed. I've
taken this as an indication that the option is not really needed,
since the value of the net.core.somaxconn sysctl or socket.SOMAXCONN
is likely to suffice.

Signed-off-by: Zac Medico <zmedico@gmail.com>